moved functions from CommunicationWrapper to lib
